update :: 30923

+++ 실버/30923.c
@@ -1,15 +1,15 @@
 #include <stdio.h>
 #include <stdlib.h>
 int main(){
-    long long int n,a,b,first;
+    int n,a,b,first;
     scanf("%lld",&n);
-    int r = n*2;
+    long long int r = n*2;
     scanf("%lld",&a);
     first = a;
     r += a*2;
     b = a;
     for(int i=1;i<n;i++){
-        scanf("%lld",&a);
+        scanf("%d",&a);
         r += a*2;
         r += abs(a-b);
         b = a;
